feat. adicao do conexao repository

+++ pw-study/pw-study/src/main/java/br/com/etechoracio/pw_study/services/DisciplinaService.java
@@ -2,6 +2,7 @@ package br.com.etechoracio.pw_study.services;
 
 import br.com.etechoracio.pw_study.entitye.Disciplina;
 import br.com.etechoracio.pw_study.repository.DisciplinaRepository;
+import br.com.etechoracio.pw_study.repository.MonitorRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -13,6 +14,9 @@ public class DisciplinaService {
 
     @Autowired
     private DisciplinaRepository repository;
+
+    @Autowired
+    private MonitorRepository monitorRepository;
 
     public List<Disciplina> listar(){
         return repository.findAll();
@@ -30,7 +34,15 @@ public class DisciplinaService {
     }
 
     public void exclude(Long id){
-        repository.deleteById(id);
+        var busca = new Disciplina();
+        busca.setId(id);
+        var monitorCadastrado = monitorRepository.findByDisciplina(busca);
+        if (monitorCadastrado.isEmpty()) {
+            repository.deleteById(id);
+        }
+        else {
+            throw new RuntimeException("Existe um monitor cadastrado a essa disciplina, não foi possível realizar a exclusão.");
+        }
     }
 
 }
